requirements.txt
appdirs==1.4.4
beautifulsoup4==4.12.3
blinker==1.8.1
bs4==0.0.2
build==1.2.1
cachelib==0.13.0
cefpython3==66.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cssselect==1.2.0
docutils==0.20.1
fake-useragent==1.5.1
Flask==3.0.3
Flask-Session==0.8.0
gunicorn==22.0.0
html5lib==1.1
idna==3.6
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.3
Kivy==2.3.0
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.7.0
Kivy-Garden==0.1.5
kivymd==1.2.0
lxml==5.2.1
MarkupSafe==2.1.5
msgspec==0.18.6
numpy==1.26.4
packaging==24.0
pandas==2.2.2
parse==1.20.1
pathlib==1.0.1
pillow==10.2.0
pyee==11.1.0
Pygments==2.17.2
pypiwin32==223
pyppeteer==2.0.0
pyproject_hooks==1.1.0
pyquery==2.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
redis==5.0.4
requests==2.31.0
requests-html==0.10.0
setuptools==69.5.1
six==1.16.0
soupsieve==2.5
spotipy==2.23.0
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==1.26.18
w3lib==2.1.2
webencodings==0.5.1
websockets==10.4
Werkzeug==3.0.2
youtube-dl==2021.12.17
zipp==3.18.1
